requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-filter==2.2.0
django-model-utils==4.0.0
django-polymorphic==2.1.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
factory-boy==2.12.0
Faker==4.0.3
idna==2.9
oauthlib==3.1.0
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
tzlocal==2.0.0
urllib3==1.25.9
